@@ -6,9 +6,10 @@ import { MainPageComponent } from './pages/main-page/main-page.component';
 import { AboutPageComponent } from './pages/about-page/about-page.component';
 import { UserPageComponent } from './pages/user-page/user-page.component';
 import { EditUserPageComponent } from './pages/edit-user-page/edit-user-page.component';
+import { UsersListPageComponent } from './pages/users-list-page/users-list-page.component';
 import { PrivacyPolicyPageComponent } from './pages/privacy-policy-page/privacy-policy-page.component';
 import { TermsOfServicePageComponent } from './pages/terms-of-service-page/terms-of-service-page.component';
-import { authGuard, loginGuard } from './guards/auth.guard';
+import { authGuard, loginGuard, adminGuard } from './guards/auth.guard';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/login', pathMatch: 'full' },
@@ -18,6 +19,7 @@ export const routes: Routes = [
   { path: 'main', component: MainPageComponent, canActivate: [authGuard] },
   { path: 'user', component: UserPageComponent, canActivate: [authGuard] },
   { path: 'edit-user', component: EditUserPageComponent, canActivate: [authGuard] },
+  { path: 'users-list', component: UsersListPageComponent, canActivate: [adminGuard] },
   { path: 'about', component: AboutPageComponent, canActivate: [authGuard] },
   { path: 'privacy-policy', component: PrivacyPolicyPageComponent, canActivate: [authGuard] },
   { path: 'terms-of-service', component: TermsOfServicePageComponent, canActivate: [authGuard] }
